Big Commit
First I changed how i was packaging the app so now there is a package
related to artist search and one related to the track list.
Then I implemented the track list feature and wired it up to the artist
activity. This is a functioning implementation but there are still a few
things to fix up before submission.